Move extractors in seperate files and add some tests
