+++ codecov.yml
@@ -1,5 +1,3 @@
-comment:
-  require_changes: true
 coverage:
   status:
     project:
