+++ src/candidate_lists/extractors/full_candidate_list.rs
@@ -1,21 +1,15 @@
 use axum::extract::{FromRef, FromRequestParts, Path};
-use serde::Deserialize;
 use sqlx::PgPool;
 
 use crate::{
     AppError, Context,
-    pagination::Pagination,
-    persons::{self, Person, PersonId, PersonPagination, PersonSort},
+    candidate_lists::{self, FullCandidateList},
     t,
 };
 
-#[derive(Deserialize)]
-struct PersonPathParams {
-    #[serde(alias = "person_id")]
-    person_id: PersonId,
-}
+use super::CandidateListPathParams;
 
-impl<S> FromRequestParts<S> for Person
+impl<S> FromRequestParts<S> for FullCandidateList
 where
     S: Send + Sync,
     PgPool: FromRef<S>,
@@ -30,52 +24,18 @@ where
         let context = Context::from_request_parts(parts, state)
             .await
             .unwrap_or_default();
-        let Path(PersonPathParams { person_id }) =
-            Path::<PersonPathParams>::from_request_parts(parts, state).await?;
+        let Path(CandidateListPathParams { list_id }) =
+            Path::<CandidateListPathParams>::from_request_parts(parts, state).await?;
 
-        let person = persons::get_person(&mut conn, person_id)
+        let full_list = candidate_lists::get_full_candidate_list(&mut conn, list_id)
             .await?
             .ok_or(AppError::NotFound(t!(
-                "person.not_found",
+                "candidate_list.not_found",
                 context.locale,
-                person_id
+                list_id
             )))?;
 
-        Ok(person)
-    }
-}
-
-impl<S> FromRequestParts<S> for PersonPagination
-where
-    S: Send + Sync,
-    PgPool: FromRef<S>,
-{
-    type Rejection = AppError;
-
-    async fn from_request_parts(
-        parts: &mut axum::http::request::Parts,
-        state: &S,
-    ) -> Result<Self, Self::Rejection> {
-        let mut conn = PgPool::from_ref(state).acquire().await?;
-        let pagination: Pagination<PersonSort> =
-            Pagination::from_request_parts(parts, state).await?;
-
-        let total_items = persons::count_persons(&mut conn).await?.max(0) as u64;
-        let pagination = pagination.set_total(total_items);
-
-        let persons = persons::list_persons(
-            &mut conn,
-            pagination.limit(),
-            pagination.offset(),
-            pagination.sort(),
-            pagination.direction(),
-        )
-        .await?;
-
-        Ok(PersonPagination {
-            persons,
-            pagination,
-        })
+        Ok(full_list)
     }
 }
 
@@ -93,27 +53,47 @@ mod tests {
     use tower::ServiceExt;
 
     use crate::{
-        AppState, Locale, persons, render_error_pages, t,
-        test_utils::{response_body_string, sample_person},
+        AppState, Locale,
+        candidate_lists::{self, CandidateListId},
+        persons::{self, PersonId},
+        render_error_pages, t,
+        test_utils::{response_body_string, sample_candidate_list, sample_person},
     };
 
     #[sqlx::test]
-    async fn person_extractor_loads_person(pool: PgPool) {
+    async fn full_candidate_list_extractor_loads_candidates(pool: PgPool) {
+        let list_id = CandidateListId::new();
+        let list = sample_candidate_list(list_id);
         let person = sample_person(PersonId::new());
+
         let mut conn = pool.acquire().await.unwrap();
+        candidate_lists::create_candidate_list(&mut conn, &list)
+            .await
+            .unwrap();
         persons::create_person(&mut conn, &person).await.unwrap();
+        candidate_lists::update_candidate_list_order(&mut conn, list_id, &[person.id])
+            .await
+            .unwrap();
 
         let app = Router::new()
             .route(
-                "/persons/{person_id}",
-                get(|person: Person| async { person.last_name }),
+                "/candidate-lists/{list_id}/full",
+                get(|full_list: FullCandidateList| async move {
+                    full_list
+                        .candidates
+                        .first()
+                        .expect("candidate")
+                        .person
+                        .last_name
+                        .clone()
+                }),
             )
             .with_state(AppState::new_for_tests(pool));
 
         let response = app
             .oneshot(
                 Request::builder()
-                    .uri(format!("/persons/{}", person.id))
+                    .uri(format!("/candidate-lists/{list_id}/full"))
                     .body(Body::empty())
                     .unwrap(),
             )
@@ -126,13 +106,13 @@ mod tests {
     }
 
     #[sqlx::test]
-    async fn person_extractor_returns_not_found(pool: PgPool) {
-        let person_id = PersonId::new();
+    async fn full_candidate_list_extractor_returns_not_found(pool: PgPool) {
+        let list_id = CandidateListId::new();
 
         let app = Router::new()
             .route(
-                "/persons/{person_id}",
-                get(|person: Person| async { person.last_name }),
+                "/candidate-lists/{list_id}/full",
+                get(|full_list: FullCandidateList| async move { full_list.list.id.to_string() }),
             )
             .layer(middleware::from_fn(render_error_pages))
             .with_state(AppState::new_for_tests(pool));
@@ -140,7 +120,7 @@ mod tests {
         let response = app
             .oneshot(
                 Request::builder()
-                    .uri(format!("/persons/{}", person_id))
+                    .uri(format!("/candidate-lists/{list_id}/full"))
                     .header(header::ACCEPT_LANGUAGE, "en")
                     .body(Body::empty())
                     .unwrap(),
@@ -150,8 +130,7 @@ mod tests {
 
         assert_eq!(response.status(), StatusCode::NOT_FOUND);
         let body = response_body_string(response).await;
-
-        let expected = t!("person.not_found", Locale::En, person_id);
+        let expected = t!("candidate_list.not_found", Locale::En, list_id);
         assert!(body.contains(&expected));
     }
 }
