+++ src/persons/structs/person.rs
@@ -76,3 +76,89 @@ impl Person {
             .unwrap_or(&["", ""])
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use sqlx::types::chrono::Utc;
+
+    fn base_person() -> Person {
+        Person {
+            id: PersonId::new(),
+            last_name: "Dijk".to_string(),
+            last_name_prefix: None,
+            initials: "A.B.".to_string(),
+            first_name: None,
+            bsn: None,
+            place_of_residence: None,
+            country_of_residence: None,
+            gender: None,
+            date_of_birth: None,
+            locality: None,
+            postal_code: None,
+            house_number: None,
+            house_number_addition: None,
+            street_name: None,
+            created_at: Utc::now(),
+            updated_at: Utc::now(),
+        }
+    }
+
+    #[test]
+    fn last_name_formats_with_optional_prefix() {
+        let mut person = base_person();
+        assert_eq!(person.last_name_with_prefix(), "Dijk");
+        assert_eq!(person.last_name_with_prefix_appended(), "Dijk");
+
+        person.last_name_prefix = Some("van".to_string());
+        assert_eq!(person.last_name_with_prefix(), "van Dijk");
+        assert_eq!(person.last_name_with_prefix_appended(), "Dijk, van");
+    }
+
+    #[test]
+    fn display_name_prefers_first_name_over_initials() {
+        let mut person = base_person();
+        person.last_name_prefix = Some("van".to_string());
+        person.first_name = Some("Anne".to_string());
+        assert_eq!(person.display_name(), "Anne van Dijk");
+
+        person.first_name = None;
+        assert_eq!(person.display_name(), "A.B. van Dijk");
+    }
+
+    #[test]
+    fn first_name_display_falls_back_to_empty_string() {
+        let mut person = base_person();
+        assert_eq!(person.first_name_display(), "");
+
+        person.first_name = Some("Henk".to_string());
+        assert_eq!(person.first_name_display(), "Henk");
+    }
+
+    #[test]
+    fn is_dutch_defaults_to_true_and_accepts_variants() {
+        let mut person = base_person();
+        assert!(person.is_dutch());
+
+        person.country_of_residence = Some("NETHERLANDS".to_string());
+        assert!(person.is_dutch());
+
+        person.country_of_residence = Some("nederland".to_string());
+        assert!(person.is_dutch());
+
+        person.country_of_residence = Some("Belgium".to_string());
+        assert!(!person.is_dutch());
+    }
+
+    #[test]
+    fn gender_key_returns_translations_or_empty_keys() {
+        let mut person = base_person();
+        assert_eq!(person.gender_key(), &["", ""]);
+
+        person.gender = Some(Gender::Male);
+        assert_eq!(person.gender_key(), t!("gender.male"));
+
+        person.gender = Some(Gender::Female);
+        assert_eq!(person.gender_key(), t!("gender.female"));
+    }
+}
